+++ components/Banner/tests/snapshot/MDCBannerSnapshotTests.m
@@ -201,6 +201,29 @@ static const CGFloat kBannerLargeContentPadding = 30.0f;
   [self generateSnapshotAndVerifyForView:self.bannerView];
 }
 
+- (void)testLongTextWithNoActionLTR {
+  // When
+  self.bannerView.textView.text = kBannerLongText;
+  self.bannerView.imageView.hidden = YES;
+  self.bannerView.leadingButton.hidden = YES;
+  self.bannerView.trailingButton.hidden = YES;
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.bannerView];
+}
+
+- (void)testLongTextWithNoActionRTL {
+  // When
+  self.bannerView.textView.text = kBannerLongText;
+  self.bannerView.imageView.hidden = YES;
+  self.bannerView.leadingButton.hidden = YES;
+  self.bannerView.trailingButton.hidden = YES;
+  [self changeViewToRTL:self.bannerView];
+
+  // Then
+  [self generateSnapshotAndVerifyForView:self.bannerView];
+}
+
 - (void)testLongTextWithSingleActionLTR {
   // When
   self.bannerView.textView.text = kBannerLongText;
@@ -436,6 +459,7 @@ static const CGFloat kBannerLargeContentPadding = 30.0f;
         [[MDCTypographyScheme alloc] initWithDefaults:MDCTypographySchemeDefaultsMaterial201902];
 
     // When
+    self.bannerView.layoutMargins = UIEdgeInsetsMake(0, 8, 0, 8);
     self.bannerView.textView.text = kBannerShortText;
     self.bannerView.textView.font = self.typographyScheme.body2;
     MDCButton *button = self.bannerView.leadingButton;
@@ -460,6 +484,7 @@ static const CGFloat kBannerLargeContentPadding = 30.0f;
     self.typographyScheme =
         [[MDCTypographyScheme alloc] initWithDefaults:MDCTypographySchemeDefaultsMaterial201902];
     MDCButton *button = self.bannerView.leadingButton;
+    self.bannerView.layoutMargins = UIEdgeInsetsMake(0, 8, 0, 8);
     [button setTitle:@"Action" forState:UIControlStateNormal];
     [button setTitleFont:self.typographyScheme.button forState:UIControlStateNormal];
     button.uppercaseTitle = YES;